ml_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# ML and Data Science
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
torch==2.1.1
transformers==4.35.2
sentence-transformers==2.2.2
prophet==1.1.4
pytorch-forecasting==0.10.4

# Clinical NLP
biopython==1.81
pubmed-lookup==0.2.1

# Drug Interaction Data
rdkit-pypi==2023.9.1

# Time Series and Forecasting
statsmodels==0.14.0
arch==6.2.0

# Monitoring and MLflow
mlflow==2.8.1
evidently==0.3.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Environment and config
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

